Version used as part of a test project to understand Orient DB
This files are experimental but working. It contains a lot of debug
prints etc both in the Java code and the Eiffel code.
The peints should be kept until the stability of the system is proven
and the API is is more definite. Then we probably have to make a more
advanced debug utility, if still needed.