Migrated to 1.7 RC2 version of Orient DB.
The API that was used in eODB was depreciated after 1.6 version. I could
have decided to stick to that version and see if that track was
developed in parallell by somebody.
However, the API I switch to, which also was supported earlier, complies
with the Tinkerpop interface which several NoSQL databases support. It
will now be possible to switch between different NoSQL databases wich
makes eODB less dependent of OrientDB

This is not a complete update but all testcases except for the root
vertex realted test pass now. The root vertex features seems not to be
supported på Tinkerpop interface. I need to investigate if I need them
and if needed how they can be implemented with this API.